WIP working CLI utils

+++ conda_smithy/feedstock_tokens.py
@@ -8,17 +8,14 @@ import git
 import requests
 import scrypt
 
-from .ci_register import (
-    circle_token,
-    drone_session,
-    travis_headers,
-    travis_get_repo_info,
-    travis_endpoint,
-)
-from .github import github_token
-
 
 def generate_and_write_feedstock_token(user, project):
+    """Generate a feedstock token and write it to
+
+        ~/.conda-smithy/{user or org}_{repo w/o '-feedstock'}_feedstock.token
+
+    This function will fail if the token file already exists.
+    """
     # capture stdout, stderr and suppress all exceptions so we don't
     # spill tokens
     failed = False
@@ -29,7 +26,7 @@ def generate_and_write_feedstock_token(user, project):
                 token = secrets.token_hex(32)
                 pth = os.path.join(
                     "~",
-                    ".conda_smithy",
+                    ".conda-smithy",
                     "%s_%s_feedstock.token" % (user, project),
                 )
                 pth = os.path.expanduser(pth)
@@ -62,19 +59,27 @@ def generate_and_write_feedstock_token(user, project):
 
 
 def read_feedstock_token(user, project):
+    """Read the feedstock token from
+
+        ~/.conda-smithy/{user or org}_{repo w/o '-feedstock'}_feedstock.token
+
+    In order to not spill any tokens to stdout/stderr, this function
+    should be used in a `try...except` block with stdout/stderr redirected
+    to /dev/null, etc.
+    """
     err_msg = None
     feedstock_token = None
 
     # read the token
     user_token_pth = os.path.join(
             "~",
-            ".conda_smithy",
+            ".conda-smithy",
             "%s_%s_feedstock.token" % (user, project),
         )
     user_token_pth = os.path.expanduser(user_token_pth)
 
     if not os.path.exists(user_token_pth):
-        err_msg = "No token found in '~/.conda_smithy/%s_%s_feedstock.token'" % (
+        err_msg = "No token found in '~/.conda-smithy/%s_%s_feedstock.token'" % (
             user,
             project,
         )
@@ -83,7 +88,7 @@ def read_feedstock_token(user, project):
             feedstock_token = fp.read().strip()
         if not feedstock_token:
             err_msg = (
-                "Empty token found in '~/.conda_smithy/"
+                "Empty token found in '~/.conda-smithy/"
                 "%s_%s_feedstock.token'"
             ) % (
                 user,
@@ -92,11 +97,89 @@ def read_feedstock_token(user, project):
     return feedstock_token, err_msg
 
 
+def feedstock_token_exists(user, project, token_repo):
+    """Test if the feedstock token exists for the given repo.
+
+    All exceptions are swallowed and stdout/stderr from this function is
+    redirected to `/dev/null`. Sanitized error messages are
+    displayed at the end.
+
+    If you need to debug this function, define `DEBUG_FEEDSTOCK_TOKENS` in
+    your environment before calling this function.
+    """
+    from .github import gh_token
+    github_token = gh_token()
+
+    exists = False
+    failed = False
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with open(os.devnull, "w") as fp:
+            with redirect_stdout(fp), redirect_stderr(fp):
+                try:
+                    feedstock_token, err_msg = read_feedstock_token(user, project)
+                    if err_msg:
+                        failed = True
+                        raise RuntimeError(err_msg)
+
+                    # clone the repo
+                    _token_repo = (
+                        token_repo
+                        .replace("$GITHUB_TOKEN", github_token)
+                        .replace("${GITHUB_TOKEN}", github_token)
+                        .replace("$GH_TOKEN", github_token)
+                        .replace("${GH_TOKEN}", github_token)
+                    )
+                    git.Repo.clone_from(_token_repo, tmpdir, depth=1)
+                    token_file = os.path.join(
+                        tmpdir,
+                        "tokens",
+                        project.replace("-feedstock", "") + ".json",
+                    )
+
+                    # don't overwrite existing tokens
+                    # check again since there might be a race condition
+                    if os.path.exists(token_file):
+                        exists = True
+                except Exception as e:
+                    if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
+                        raise e
+                    failed = True
+    if failed:
+        if err_msg:
+            raise RuntimeError(err_msg)
+        else:
+            raise RuntimeError(
+                (
+                    "Testing for the feedstock token for %s/%s failed!"
+                    " Try the command locally with DEBUG_FEEDSTOCK_TOKENS"
+                    " defined in the environment to investigate!") % (user, project)
+                )
+
+    return exists
+
+
 def register_feedstock_token(user, project, token_repo):
-    # capture stdout, stderr and suppress all exceptions so we don't
-    # spill tokens
+    """Register the feedstock token with the token repo.
+
+    This function uses a random salt and scrypt to hash the feedstock
+    token before writing it to the token repo. NEVER STORE THESE TOKENS
+    IN PLAIN TEXT!
+
+    All exceptions are swallowed and stdout/stderr from this function is
+    redirected to `/dev/null`. Sanitized error messages are
+    displayed at the end.
+
+    If you need to debug this function, define `DEBUG_FEEDSTOCK_TOKENS` in
+    your environment before calling this function.
+    """
+    from .github import gh_token
+    github_token = gh_token()
+
     failed = False
     err_msg = None
+
+    # capture stdout, stderr and suppress all exceptions so we don't
+    # spill tokens
     with tempfile.TemporaryDirectory() as tmpdir:
         with open(os.devnull, "w") as fp:
             with redirect_stdout(fp), redirect_stderr(fp):
@@ -117,10 +200,12 @@ def register_feedstock_token(user, project, token_repo):
                     repo = git.Repo.clone_from(_token_repo, tmpdir, depth=1)
                     token_file = os.path.join(
                         tmpdir,
+                        "tokens",
                         project.replace("-feedstock", "") + ".json",
                     )
 
                     # don't overwrite existing tokens
+                    # check again since there might be a race condition
                     if os.path.exists(token_file):
                         failed = True
                         err_msg = "Token for repo %s/%s already exists!" % (user, project)
@@ -135,12 +220,12 @@ def register_feedstock_token(user, project, token_repo):
                         maxtime=float(maxtime),
                     )
                     data = {
-                        "salt": salt,
-                        "encrypted_token": salted_token,
+                        "salt": salt.hex(),
+                        "hashed_token": salted_token.hex(),
                         "maxtime": maxtime,
                     }
                     with open(token_file, "w") as fp:
-                        fp.write(json.dump(data))
+                        json.dump(data, fp)
 
                     # push
                     repo.index.add(token_file)
@@ -168,8 +253,26 @@ def register_feedstock_token(user, project, token_repo):
 def register_feedstock_token_with_proviers(
         user, project, feedstock_directory,
         drone=True, circle=True,
-        travis=True, azure=True
+        travis=True, azure=True,
+        clobber=True,
 ):
+    """Register the feedstock token with provider CI services.
+
+    Note that if a feedstock token is already registered and `clobber=True`
+    this function will overwrite existing tokens.
+
+    All exceptions are swallowed and stdout/stderr from this function is
+    redirected to `/dev/null`. Sanitized error messages are
+    displayed at the end.
+
+    If you need to debug this function, define `DEBUG_FEEDSTOCK_TOKENS` in
+    your environment before calling this function.
+    """
+    # we are swallong all of the logs below, so we do a test import here
+    # to generate the proper errors for missing tokens
+    from .ci_register import travis_endpoint  # noqa
+    from .azure_ci_utils import default_config  # noqa
+
     # capture stdout, stderr and suppress all exceptions so we don't
     # spill tokens
     failed = False
@@ -184,7 +287,7 @@ def register_feedstock_token_with_proviers(
 
                 if circle:
                     try:
-                        add_feedstock_token_to_circle(user, project, feedstock_token)
+                        add_feedstock_token_to_circle(user, project, feedstock_token, clobber)
                     except Exception as e:
                         if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
                             raise e
@@ -197,7 +300,7 @@ def register_feedstock_token_with_proviers(
 
                 if drone:
                     try:
-                        add_feedstock_token_to_drone(user, project, feedstock_token)
+                        add_feedstock_token_to_drone(user, project, feedstock_token, clobber)
                     except Exception as e:
                         if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
                             raise e
@@ -210,7 +313,7 @@ def register_feedstock_token_with_proviers(
 
                 if travis:
                     try:
-                        add_feedstock_token_to_travis(user, project, feedstock_token)
+                        add_feedstock_token_to_travis(user, project, feedstock_token, clobber)
                     except Exception as e:
                         if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
                             raise e
@@ -223,7 +326,7 @@ def register_feedstock_token_with_proviers(
 
                 if azure:
                     try:
-                        add_feedstock_token_to_azure(user, project, feedstock_token)
+                        add_feedstock_token_to_azure(user, project, feedstock_token, clobber)
                     except Exception as e:
                         if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
                             raise e
@@ -250,7 +353,9 @@ def register_feedstock_token_with_proviers(
                 )
 
 
-def add_feedstock_token_to_circle(user, project, feedstock_token):
+def add_feedstock_token_to_circle(user, project, feedstock_token, clobber):
+    from .ci_register import circle_token
+
     url_template = (
         "https://circleci.com/api/v1.1/project/github/{user}/{project}/envvar{extra}?"
         "circle-token={token}"
@@ -265,7 +370,7 @@ def add_feedstock_token_to_circle(user, project, feedstock_token):
         if evar["name"] == "FEEDSTOCK_TOKEN":
             have_feedstock_token = True
 
-    if have_feedstock_token:
+    if have_feedstock_token and clobber:
         r = requests.delete(url_template.format(
             token=circle_token,
             user=user,
@@ -275,16 +380,19 @@ def add_feedstock_token_to_circle(user, project, feedstock_token):
         if r.status_code != 200:
             r.raise_for_status()
 
-    data = {"name": "FEEDSTOCK_TOKEN", "value": feedstock_token}
-    response = requests.post(
-        url_template.format(token=circle_token, user=user, project=project, extra=""),
-        data,
-    )
-    if response.status_code != 201:
-        raise ValueError(response)
+    if not have_feedstock_token or (have_feedstock_token and clobber):
+        data = {"name": "FEEDSTOCK_TOKEN", "value": feedstock_token}
+        response = requests.post(
+            url_template.format(token=circle_token, user=user, project=project, extra=""),
+            data,
+        )
+        if response.status_code != 201:
+            raise ValueError(response)
 
 
-def add_feedstock_token_to_drone(user, project, feedstock_token):
+def add_feedstock_token_to_drone(user, project, feedstock_token, clobber):
+    from .ci_register import drone_session
+
     session = drone_session()
 
     r = session.get(f"/api/repos/{user}/{project}/secrets")
@@ -294,7 +402,7 @@ def add_feedstock_token_to_drone(user, project, feedstock_token):
         if "FEEDSTOCK_TOKEN" == secret["name"]:
             have_feedstock_token = True
 
-    if have_feedstock_token:
+    if have_feedstock_token and clobber:
         r = session.patch(
             f"/api/repos/{user}/{project}/secrets/FEEDSTOCK_TOKEN",
             json={
@@ -303,7 +411,7 @@ def add_feedstock_token_to_drone(user, project, feedstock_token):
             },
         )
         r.raise_for_status()
-    else:
+    elif not have_feedstock_token:
         response = session.post(
             f"/api/repos/{user}/{project}/secrets",
             json={
@@ -316,8 +424,9 @@ def add_feedstock_token_to_drone(user, project, feedstock_token):
             response.raise_for_status()
 
 
-def add_feedstock_token_to_travis(user, project, feedstock_token):
+def add_feedstock_token_to_travis(user, project, feedstock_token, clobber):
     """Add the FEEDSTOCK_TOKEN to travis."""
+    from .ci_register import travis_endpoint, travis_headers, travis_get_repo_info
 
     headers = travis_headers()
 
@@ -344,7 +453,7 @@ def add_feedstock_token_to_travis(user, project, feedstock_token):
         "env_var.public": "false",
     }
 
-    if have_feedstock_token:
+    if have_feedstock_token and clobber:
         r = requests.patch(
             "{}/repo/{repo_id}/env_var/{ev_id}".format(
                 travis_endpoint,
@@ -355,7 +464,7 @@ def add_feedstock_token_to_travis(user, project, feedstock_token):
             json=data,
         )
         r.raise_for_status()
-    else:
+    elif not have_feedstock_token:
         r = requests.post(
             "{}/repo/{repo_id}/env_vars".format(travis_endpoint, repo_id=repo_id),
             headers=headers,
@@ -365,7 +474,7 @@ def add_feedstock_token_to_travis(user, project, feedstock_token):
             r.raise_for_status()
 
 
-def add_feedstock_token_to_azure(user, project, feedstock_token):
+def add_feedstock_token_to_azure(user, project, feedstock_token, clobber):
     from .azure_ci_utils import build_client
     from .azure_ci_utils import default_config as config
     from vsts.build.v4_1.models import BuildDefinitionVariable
@@ -383,13 +492,19 @@ def add_feedstock_token_to_azure(user, project, feedstock_token):
             "Cannot add FEEDSTOCK_TOKEN to a repo that is not already registerd on azure CI!"
         )
 
-    if ed.variables is None:
+    if not hasattr(ed, "variables") or ed.variables is None:
         ed.variables = {}
 
-    ed.variables["FEEDSTOCK_TOKEN"] = BuildDefinitionVariable(
-        allow_override=False, is_secret=True, value=feedstock_token,
-    )
+    if "FEEDSTOCK_TOKEN" in ed.variables:
+        have_feedstock_token = True
+    else:
+        have_feedstock_token = False
 
-    bclient.update_definition(
-        definition=ed, definition_id=ed.id, project=ed.project.name,
-    )
+    if not have_feedstock_token or (have_feedstock_token and clobber):
+        ed.variables["FEEDSTOCK_TOKEN"] = BuildDefinitionVariable(
+            allow_override=False, is_secret=True, value=feedstock_token,
+        )
+
+        bclient.update_definition(
+            definition=ed, definition_id=ed.id, project=ed.project.name,
+        )
